pyproject.toml
[tool.poetry]
name = "grocery-api"
version = "0.1.0"
description = ""
authors = ["Zach Wright <zachw@gurutechnologies.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
django-nested-admin = "*"
django-filter = "^24.2"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
